Permissions

+++ collabl/collaborations/views_htmx.py
@@ -2,10 +2,9 @@ from datetime import datetime
 
 from django.contrib import messages
 from django.contrib.auth.decorators import login_required
-from django.http import HttpResponseRedirect
+from django.http import HttpResponseRedirect, HttpResponseForbidden
 from django.shortcuts import render, get_object_or_404
 from django.urls import reverse_lazy
-from django.views.generic.list import BaseListView
 
 import collaborations.constants as c
 from collaborations.forms import (
@@ -26,7 +25,11 @@ from collaborations.utils import get_all_elements
 from collabl.settings import SITE_PROTOCOL, SITE_DOMAIN
 from groups.constants import MEMBERSHIP_STATUS_ADMIN
 from groups.models import Group
-from groups.utils import get_filtered_collaborations
+from groups.utils import (
+    get_filtered_collaborations,
+    user_is_admin,
+    user_has_active_membership,
+)
 
 
 @login_required()
@@ -37,16 +40,25 @@ def group_collaboration_create_view(request, slug):
     If "collaboration_creation_modal": True is in the context (and the form), a modal will be rendered (with error messages, if appropriate)
     """
 
+    # Get Data
     form = CollaborationForm(request.POST or None)
     group = get_object_or_404(Group, slug=slug)
 
+    # Check permissions
+    if not user_is_admin(request.user, group):
+        return HttpResponseForbidden()
+
+    # If POST, process the creation
     if request.method == "POST":
         if form.is_valid():
+            # create the collaboration
             collaboration = form.save(commit=False)
             collaboration.related_group = group
             collaboration.created_by = request.user
             collaboration.created_at = datetime.now()
             collaboration.save()
+
+            # Get success_url - we send back a javascript redirect to take the user to this page
             success_url = (
                 SITE_PROTOCOL
                 + SITE_DOMAIN
@@ -57,6 +69,8 @@ def group_collaboration_create_view(request, slug):
             )
 
             # Get filter parameter - if not set, send back a 'hidden' response (Empty HTML string)
+            # NOTE: If we are sure we wnt to redirect the user, we can cut a lot of this out, as
+            # no update to state is needed.
             collaboration_list_filter = request.GET.get(
                 "collaboration_list_filter", None
             )
@@ -65,6 +79,7 @@ def group_collaboration_create_view(request, slug):
                     "collaboration_list_filter", "HIDE"
                 )
 
+            # Return the response (with a redirect URL)
             return render(
                 request,
                 "app/group/partials/collaborations/list.html",
@@ -79,6 +94,7 @@ def group_collaboration_create_view(request, slug):
                 },
             )
 
+    # If GET, (or invalid data is posted) send back the 'Create Collaboration' Modal
     return render(
         request,
         "app/group/partials/modals/collaboration_create.html",
@@ -98,19 +114,25 @@ def collaboration_delete_view(request, slug):
     If "task_delete_modal": True is in the context, a modal will be rendered
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
 
+    # Check permissions
+    if not user_is_admin(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, process the delete operation
     if request.method == "POST":
-        group = collaboration.related_group
         collaboration.delete()
         messages.success(request, "Collaboration Removed")
         return HttpResponseRedirect(
             reverse_lazy(
                 "group-detail",
-                kwargs={"slug": group.slug},
+                kwargs={"slug": collaboration.related_group.slug},
             )
         )
 
+    # If GET, send back the 'Delete Collaboration' Modal
     return render(
         request,
         "app/collaborations/partials/header/modals/collaboration_delete.html",
@@ -129,10 +151,15 @@ def collaboration_update_view(request, slug):
     messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
-
     form = CollaborationForm(request.POST or None, instance=collaboration)
 
+    # Check permissions
+    if not user_is_admin(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, update the collaboration
     if request.method == "POST":
         if form.is_valid():
             form.save()
@@ -144,6 +171,7 @@ def collaboration_update_view(request, slug):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the populated 'Update Collaboration' Modal
     return render(
         request,
         "app/collaborations/partials/header/modals/collaboration_update.html",
@@ -163,12 +191,17 @@ def collaboration_image_view(request, slug):
     messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
-
     form = CollaborationImageForm(
         request.POST or None, request.FILES or None, instance=collaboration
     )
 
+    # Check permissions
+    if not user_is_admin(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, update the collaboration image
     if request.method == "POST":
         if form.is_valid():
             form.save()
@@ -180,6 +213,7 @@ def collaboration_image_view(request, slug):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the populated 'Collaboration Image' Modal
     return render(
         request,
         "app/collaborations/partials/header/modals/collaboration_image_update.html",
@@ -198,9 +232,15 @@ def collaboration_task_create_view(request, slug):
     If "task_creation_modal": True is in the context (and the form), a modal will be rendered (with error messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
     form = TaskForm(request.POST or None, initial={"collaboration": collaboration})
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, create the task and update the state of the collaboration
     if request.method == "POST" and form.is_valid():
         task = form.save(commit=False)
         task.collaboration = collaboration
@@ -215,6 +255,7 @@ def collaboration_task_create_view(request, slug):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Creation Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -235,13 +276,18 @@ def collaboration_task_update_view(request, slug, pk):
     If "task_update_modal": True is in the context (and the form), a modal will be rendered (with error messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
     task = get_object_or_404(CollaborationTask, pk=pk)
-
     form = TaskUpdateForm(
         request.POST or None, initial={"collaboration": collaboration}, instance=task
     )
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, update the task and then the state of the collaboration
     if request.method == "POST" and form.is_valid():
         form.save()
         return render(
@@ -253,6 +299,7 @@ def collaboration_task_update_view(request, slug, pk):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Update Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -275,11 +322,16 @@ def collaboration_task_notes_view(request, slug, pk):
     rendered (with error messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
     task = get_object_or_404(CollaborationTask, pk=pk)
-
     form = TaskCompleteForm(request.POST, request.FILES, instance=task)
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, update the task and then the state of the collaboration
     if request.method == "POST" and form.is_valid():
         form.save()
         return render(
@@ -291,6 +343,7 @@ def collaboration_task_notes_view(request, slug, pk):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -310,10 +363,15 @@ def collaboration_task_toggle_view(request, slug, pk, status):
     HTMX VIEW - Allows completion of tasks with one click and no reload
     """
 
+    # Get Data
     task = get_object_or_404(CollaborationTask, pk=pk)
-
     collaboration = task.collaboration
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # Process the request
     match status:
         case c.COMPLETE_TASK:
             task.completed_at = datetime.now()
@@ -328,6 +386,7 @@ def collaboration_task_toggle_view(request, slug, pk, status):
         case _:
             pass
 
+    # Update the state of the elements list on the collaboration page
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -351,9 +410,15 @@ def collaboration_task_delete_view(request, slug, pk):
     If "task_delete_modal": True is in the context, a modal will be rendered
     """
 
-    collaboration = get_object_or_404(Collaboration, slug=slug)
+    # Get Data
     task = get_object_or_404(CollaborationTask, pk=pk)
+    collaboration = task.collaboration
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, delete the task and then update the state of the collaboration
     if request.method == "POST":
         task.remove()
         return render(
@@ -365,6 +430,7 @@ def collaboration_task_delete_view(request, slug, pk):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -385,9 +451,15 @@ def collaboration_milestone_create_view(request, slug):
     If "milestone_creation_modal": True is in the context (and the form), a modal will be rendered (with error messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
     form = MilestoneForm(request.POST or None)
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, process the creation
     if request.method == "POST" and form.is_valid():
         milestone = form.save(commit=False)
         milestone.collaboration = collaboration
@@ -402,6 +474,7 @@ def collaboration_milestone_create_view(request, slug):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the 'Create Milestone' Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -422,15 +495,20 @@ def collaboration_milestone_update_view(request, slug, pk):
     If "task_update_modal": True is in the context (and the form), a modal will be rendered (with error messages, if appropriate)
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
     milestone = get_object_or_404(CollaborationMilestone, pk=pk)
-
     form = MilestoneForm(
         request.POST or None,
         initial={"collaboration": collaboration},
         instance=milestone,
     )
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, update the milestone and then the state of the collaboration
     if request.method == "POST" and form.is_valid():
         form.save()
         return render(
@@ -442,6 +520,7 @@ def collaboration_milestone_update_view(request, slug, pk):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Update Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -463,9 +542,15 @@ def collaboration_milestone_delete_view(request, slug, pk):
     If "task_delete_modal": True is in the context, a modal will be rendered
     """
 
+    # Get Data
     collaboration = get_object_or_404(Collaboration, slug=slug)
     milestone = get_object_or_404(CollaborationMilestone, pk=pk)
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # If POST, delete the milestone and then update the state of the collaboration
     if request.method == "POST":
         milestone.remove()
         return render(
@@ -477,6 +562,7 @@ def collaboration_milestone_delete_view(request, slug, pk):
             },
         )
 
+    # If GET, (or invalid data is posted) send back the Modal
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -493,10 +579,17 @@ def collaboration_milestone_delete_view(request, slug, pk):
 def collaboration_elements_list_view(request, slug):
     """
     HTMX VIEW - Sends back html list of elements
+    Used to refresh state after reordering takes place
     """
 
+    # Get data
     collaboration = get_object_or_404(Collaboration, slug=slug)
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # Make Response
     return render(
         request,
         "app/collaborations/partials/elements/list/main.html",
@@ -513,12 +606,20 @@ def collaboration_task_move_view(request, slug, pk, position):
     HTMX VIEW - Allows reordering of tasks.
     """
 
+    # Get data
+    collaboration = get_object_or_404(Collaboration, slug=slug)
     task = get_object_or_404(CollaborationTask, pk=pk)
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # Get the object and process the request (much of the logic is stored in the model)
     if 0 <= int(position) < task.collaboration.number_of_elements:
         task.position = int(position)
         task.save()
 
+    # Make Response
     return HttpResponseRedirect(
         reverse_lazy(
             "collaboration-elements-list",
@@ -533,54 +634,25 @@ def collaboration_milestone_move_view(request, slug, pk, position):
     HTMX VIEW - Allows reordering of milestones
     """
 
+    # Get data
+    collaboration = get_object_or_404(Collaboration, slug=slug)
     milestone = get_object_or_404(CollaborationMilestone, pk=pk)
 
+    # Check permissions
+    if not user_has_active_membership(request.user, collaboration.related_group):
+        return HttpResponseForbidden()
+
+    # Get the object and process the request (much of the logic is stored in the model)
     if 0 <= int(position) < milestone.collaboration.number_of_elements:
         milestone.position = int(position)
         milestone.save()
 
+    # Make Response
     return HttpResponseRedirect(
         reverse_lazy(
             "collaboration-elements-list",
             kwargs={"slug": milestone.collaboration.slug},
         )
-    )
-
-
-@login_required()
-def user_collaboration_create_view(request, slug):
-    """
-    HTMX VIEW - Allows milestone creation with update and no reload
-    Sends back a list of elements, to replace the content in #element_list
-    If "milestone_creation_modal": True is in the context (and the form), a modal will be rendered (with error messages, if appropriate)
-    """
-
-    collaboration = get_object_or_404(Collaboration, slug=slug)
-    form = MilestoneForm(request.POST or None)
-
-    if request.method == "POST" and form.is_valid():
-        milestone = form.save(commit=False)
-        milestone.collaboration = collaboration
-        milestone.save()
-
-        return render(
-            request,
-            "app/collaborations/partials/elements/list/main.html",
-            {
-                "elements": get_all_elements(collaboration),
-                "collaboration": collaboration,
-            },
-        )
-
-    return render(
-        request,
-        "app/collaborations/partials/elements/list/main.html",
-        {
-            "elements": get_all_elements(collaboration),
-            "collaboration": collaboration,
-            "milestone_creation_modal": True,
-            "form": form,
-        },
     )
 
 
@@ -594,15 +666,28 @@ def user_collaboration_create_view(request):
     on successful submission, sends back a JS/htmx redirect to the new group page.
     """
 
+    # Check the admin is a member of at least one group
     if not request.user.memberships.filter(status=MEMBERSHIP_STATUS_ADMIN):
         return render(request, "app/home/modals/join_or_make_a_group.html", {})
 
+    # Get data
     form = CollaborationCreateFormWithGroupSelection(request.user, request.POST or None)
+
+    # If POST, process the request, and redirect to the new collaboration
     if request.method == "POST":
         if form.is_valid():
+
+            # Check permissions
+            group = form.instance.related_group
+            if not user_is_admin(request.user, group):
+                return HttpResponseForbidden()
+
+            # Create Collaboration
             collaboration = form.save(commit=False)
             collaboration.created_by = request.user
             collaboration.save()
+
+            # Get success_url - we send back a javascript redirect to take the user to this page
             success_url = (
                 SITE_PROTOCOL
                 + SITE_DOMAIN
@@ -612,6 +697,7 @@ def user_collaboration_create_view(request):
                 )
             )
 
+            # Make Response (redirect)
             return render(
                 request,
                 "app/snippets/js_redirect.html",
@@ -620,6 +706,7 @@ def user_collaboration_create_view(request):
                 },
             )
 
+    # Make Response (Creation Modal)
     return render(
         request,
         "app/home/modals/collaboration_create.html",
